chore: extract NodeAuthTab for reus
