+++ packages/shared/src/components/molecules/index.ts
@@ -8,6 +8,7 @@ export { default as NetworkRecipientSelector } from './NetworkRecipientSelector.
 export { default as NftGallery } from './NftGallery.svelte'
 export { default as NftGalleryItem } from './NftGalleryItem.svelte'
 export { default as NftMedia } from './NftMedia.svelte'
+export { default as NodeAuthTab } from './NodeAuthTab.svelte'
 export { default as SelectorInput } from './SelectorInput.svelte'
 export { default as ShimmerClaimingAccountList } from './ShimmerClaimingAccountList.svelte'
 export { default as TransactionAssetSection } from './TransactionAssetSection.svelte'
